pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "jornal-agent"
version = "1.0.0"
description = "Agente automatizado para download, processamento e envio de clipagem de jornal"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jornal-Agent Team"}
]
keywords = ["automation", "pdf", "ocr", "whatsapp", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: Office/Business :: News/Diary",
]

dependencies = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyMuPDF>=1.23.0",
    "pdf2image>=1.16.3",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "openai>=1.6.0",
    "twilio>=8.10.0",
    "loguru>=0.7.2",
    "tenacity>=8.2.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
jornal-agent = "src.main:main"

[project.urls]
Homepage = "https://github.com/lenondpaula/goodluke"
Repository = "https://github.com/lenondpaula/goodluke"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
